Split into playlist.js and video.js
